Working verion with early NetOptimization Needs work
