+++ computer/10/buffer.txt
@@ -1,401 +1,58 @@
 {
   {
     y = 1,
-    x = 28,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.053333282471,
-  },
-  {
-    y = 1,
-    x = 27,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.053055763245,
-  },
-  {
-    y = 1,
-    x = 26,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.052778244019,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.052499771118,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 23,
-    timestamp = 14.052222251892,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 22,
-    timestamp = 14.051944732666,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 21,
-    timestamp = 14.05138874054,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.051111221313,
-  },
-  {
-    y = 1,
-    x = 26,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.050833702087,
-  },
-  {
-    y = 1,
-    x = 27,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.050556182861,
-  },
-  {
-    y = 1,
-    x = 28,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.050277709961,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.049722671509,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 20,
-    z = 21,
-    timestamp = 14.049445152283,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 20,
-    z = 22,
-    timestamp = 14.049166679382,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 20,
-    z = 23,
-    timestamp = 14.048889160156,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.04861164093,
-  },
-  {
-    y = 1,
-    x = 28,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.048334121704,
-  },
-  {
-    y = 1,
-    x = 27,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.048055648804,
-  },
-  {
-    y = 1,
-    x = 26,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.047778129578,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.047223091125,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 23,
-    timestamp = 14.046944618225,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 19,
-    z = 20,
-    timestamp = 14.046667098999,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 22,
-    timestamp = 14.046667098999,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 21,
-    timestamp = 14.046389579773,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 19,
-    z = 19,
-    timestamp = 14.046389579773,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 19,
-    z = 18,
-    timestamp = 14.046111106873,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.046111106873,
-  },
-  {
-    y = 1,
-    x = 26,
-    turtleId = 19,
-    z = 18,
-    timestamp = 14.045833587646,
-  },
-  {
-    y = 1,
-    x = 26,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.045833587646,
-  },
-  {
-    y = 1,
-    x = 27,
-    turtleId = 19,
-    z = 18,
-    timestamp = 14.04555606842,
-  },
-  {
-    y = 1,
-    x = 27,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.04555606842,
-  },
-  {
-    y = 1,
-    x = 28,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.045000076294,
-  },
-  {
-    y = 1,
-    x = 28,
-    turtleId = 19,
-    z = 18,
-    timestamp = 14.045000076294,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 20,
-    z = 20,
-    timestamp = 14.044722557068,
-  },
-  {
-    y = 1,
     x = 29,
     turtleId = 19,
     z = 18,
-    timestamp = 14.044722557068,
+    timestamp = 16.879722595215,
   },
   {
     y = 1,
-    x = 29,
-    turtleId = 19,
-    z = 19,
-    timestamp = 14.044445037842,
-  },
-  {
-    y = 1,
-    x = 29,
+    x = 32,
     turtleId = 20,
-    z = 21,
-    timestamp = 14.044445037842,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 20,
-    z = 22,
-    timestamp = 14.043889045715,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 19,
     z = 20,
-    timestamp = 14.043889045715,
+    timestamp = 16.879722595215,
   },
   {
     y = 1,
-    x = 29,
+    x = 33,
     turtleId = 20,
-    z = 23,
-    timestamp = 14.043611526489,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 19,
-    z = 21,
-    timestamp = 14.043611526489,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.043055534363,
-  },
-  {
-    y = 1,
-    x = 29,
-    turtleId = 19,
-    z = 22,
-    timestamp = 14.043055534363,
-  },
-  {
-    y = 1,
-    x = 28,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.042778015137,
-  },
-  {
-    y = 1,
-    x = 28,
-    turtleId = 19,
-    z = 22,
-    timestamp = 14.042778015137,
-  },
-  {
-    y = 1,
-    x = 27,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.042222976685,
-  },
-  {
-    y = 1,
-    x = 27,
-    turtleId = 19,
-    z = 22,
-    timestamp = 14.042222976685,
-  },
-  {
-    y = 1,
-    x = 26,
-    turtleId = 20,
-    z = 24,
-    timestamp = 14.041944503784,
-  },
-  {
-    y = 1,
-    x = 26,
-    turtleId = 19,
-    z = 22,
-    timestamp = 14.041944503784,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 19,
-    z = 22,
-    timestamp = 14.041389465332,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 19,
-    z = 21,
-    timestamp = 14.041111946106,
-  },
-  {
-    y = 1,
-    x = 25,
-    turtleId = 19,
     z = 20,
-    timestamp = 14.040555953979,
+    timestamp = 16.879444122314,
   },
   {
     y = 1,
-    x = 25,
-    turtleId = 19,
-    z = 19,
-    timestamp = 14.040278434753,
-  },
-  {
-    y = 1,
-    x = 25,
+    x = 30,
     turtleId = 19,
     z = 18,
-    timestamp = 14.039722442627,
+    timestamp = 16.879444122314,
   },
   {
     y = 1,
-    x = 26,
+    x = 31,
     turtleId = 19,
     z = 18,
-    timestamp = 14.039444923401,
+    timestamp = 16.879167556763,
   },
   {
     y = 1,
-    x = 27,
-    turtleId = 19,
-    z = 18,
-    timestamp = 14.038888931274,
+    x = 34,
+    turtleId = 20,
+    z = 20,
+    timestamp = 16.879167556763,
   },
   {
     y = 1,
-    x = 28,
+    x = 32,
     turtleId = 19,
     z = 18,
-    timestamp = 14.038611412048,
+    timestamp = 16.878889083862,
+  },
+  {
+    y = 1,
+    x = 35,
+    turtleId = 20,
+    z = 20,
+    timestamp = 16.878889083862,
   },
 }
