--- computer/10/bufferI.txt
@@ -1,7 +1,0 @@
-"{\
-  y = 1,\
-  x = 28,\
-  turtleId = 19,\
-  z = 18,\
-  timestamp = 14.038611412048,\
-}"
