--- computer/10/e.txt
@@ -1,254 +1,0 @@
-{
-  "modem_message",
-  "right",
-  43,
-  15,
-  "{\
-  {\
-    y = 1,\
-    x = 26,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.041944503784,\
-  },\
-  {\
-    y = 1,\
-    x = 27,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.042222976685,\
-  },\
-  {\
-    y = 1,\
-    x = 28,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.042778015137,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.043055534363,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 23,\
-    timestamp = 14.043611526489,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 22,\
-    timestamp = 14.043889045715,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 21,\
-    timestamp = 14.044445037842,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.044722557068,\
-  },\
-  {\
-    y = 1,\
-    x = 28,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.045000076294,\
-  },\
-  {\
-    y = 1,\
-    x = 27,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.04555606842,\
-  },\
-  {\
-    y = 1,\
-    x = 26,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.045833587646,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.046111106873,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 21,\
-    timestamp = 14.046389579773,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 22,\
-    timestamp = 14.046667098999,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 23,\
-    timestamp = 14.046944618225,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.047223091125,\
-  },\
-  {\
-    y = 1,\
-    x = 26,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.047778129578,\
-  },\
-  {\
-    y = 1,\
-    x = 27,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.048055648804,\
-  },\
-  {\
-    y = 1,\
-    x = 28,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.048334121704,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.04861164093,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 23,\
-    timestamp = 14.048889160156,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 22,\
-    timestamp = 14.049166679382,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 21,\
-    timestamp = 14.049445152283,\
-  },\
-  {\
-    y = 1,\
-    x = 29,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.049722671509,\
-  },\
-  {\
-    y = 1,\
-    x = 28,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.050277709961,\
-  },\
-  {\
-    y = 1,\
-    x = 27,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.050556182861,\
-  },\
-  {\
-    y = 1,\
-    x = 26,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.050833702087,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 20,\
-    timestamp = 14.051111221313,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 21,\
-    timestamp = 14.05138874054,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 22,\
-    timestamp = 14.051944732666,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 23,\
-    timestamp = 14.052222251892,\
-  },\
-  {\
-    y = 1,\
-    x = 25,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.052499771118,\
-  },\
-  {\
-    y = 1,\
-    x = 26,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.052778244019,\
-  },\
-  {\
-    y = 1,\
-    x = 27,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.053055763245,\
-  },\
-  {\
-    y = 1,\
-    x = 28,\
-    turtleId = 20,\
-    z = 24,\
-    timestamp = 14.053333282471,\
-  },\
-}",
-  29.359836511806,
-}
