+++ computer/10/packets.txt
@@ -1,0 +1,30 @@
+{
+  {
+    y = 1,
+    x = 32,
+    turtleId = 19,
+    z = 18,
+    timestamp = 16.878889083862,
+  },
+  {
+    y = 1,
+    x = 31,
+    turtleId = 19,
+    z = 18,
+    timestamp = 16.879167556763,
+  },
+  {
+    y = 1,
+    x = 30,
+    turtleId = 19,
+    z = 18,
+    timestamp = 16.879444122314,
+  },
+  {
+    y = 1,
+    x = 29,
+    turtleId = 19,
+    z = 18,
+    timestamp = 16.879722595215,
+  },
+}
