--- computer/10/packetsJ.txt
@@ -1,7 +1,0 @@
-{
-  y = 1,
-  x = 28,
-  turtleId = 20,
-  z = 24,
-  timestamp = 14.053333282471,
-}
