--- computer/15/out.txt
@@ -1,8 +1,0 @@
-{
-  time = "Sun Jun 29 14:27:23 2025",
-  coords = {
-    y = 2,
-    x = 1,
-    z = 3,
-  },
-}
