+++ computer/20/startup.lua
@@ -5,19 +5,33 @@ local modem = peripheral.find("modem") or error("No modem attached", 0)
 local RECEIVE_CHANNEL = 15
 local SENDING_CHANNEL = 43
 modem.open(RECEIVE_CHANNEL)
+function restart()
+    print("rebooting..")
+    sleep(2)
+    os.reboot()
+end
+
 function MovementLoop()
     print("thread 1")
     while true do
         -- max 55=x, 32=z/y because we are only looking at a one y coord at a time
         
         sleep(1)
-        currentMove = TMNL.Forward()
+        TMNL.Forward()
+         sleep(1)
+        TMNL.Forward()
         sleep(1)
-        currentMove = TMNL.Forward()
+        TMNL.Forward()
         sleep(1)
-        currentMove = TMNL.Forward()
+        TMNL.Forward()
         sleep(1)
-        currentMove = TMNL.Forward()
+        TMNL.Forward()
+        sleep(1)
+        TMNL.Forward()
+        sleep(1)
+        TMNL.Forward()
+        sleep(1)
+        TMNL.Forward()
         
         TMNL.TurnLeft()
     end
@@ -27,20 +41,25 @@ function ListenLoop()
     while true do
         local e = { os.pullEvent() }
         if (e[1] == "modem_message") then
+            print("EVENT: " .. textutils.serialize(e))
+
             pack = e[5]
-            print("EVENT: " .. textutils.serialize(e[5]))
-            if (pack == "send latest2") then
-                print("sending Packet")
-                sleep(.7)
-                modem.transmit(SENDING_CHANNEL,RECEIVE_CHANNEL,textutils.serialize(TMNL.Queue))
-                TMNL.Queue = {}
-            end
+            if (pack == "send latest2") then    
+                print("EVENT: " .. textutils.serialize(e[5]))
+                if (#TMNL.Queue > 0) then
+                    print("sending Packet")
+                    modem.transmit(SENDING_CHANNEL,RECEIVE_CHANNEL,textutils.serialize(TMNL.Queue))
+                    TMNL.Queue = {}
+                else
+                    print("empty queue")         
+                end
+            else
+                print("wrong pack")
+             end
         end
     end
 end
-print("please enter y to begin mining: ")
-userInput = read()
-if (userInput == 'y') then
+
     print("Happy Mining")
     print("waiting")
     C,RC,Message,D = MineNet.listenOnChannel(RECEIVE_CHANNEL)
@@ -55,12 +74,11 @@ if (userInput == 'y') then
             parallel.waitForAny(MovementLoop,ListenLoop)
         else
             print("begin mining not recieved")
+            restart()
         end
     else
         print("hello was not recieved")
+        restart()
     end
 
     
-else
-    print("invalid") 
-end
