+++ computer/6/e.txt
@@ -5,10 +5,10 @@
   73,\
   \"{\\\
   y = 1,\\\
-  x = 28,\\\
-  turtleId = 19,\\\
-  z = 18,\\\
-  timestamp = 14.038611412048,\\\
+  x = 35,\\\
+  turtleId = 20,\\\
+  z = 20,\\\
+  timestamp = 16.878889083862,\\\
 }\",\
   5.4772255750517,\
 }"
