+++ computer/6/e5.txt
@@ -1,7 +1,7 @@
 "\"{\\\
   y = 1,\\\
-  x = 28,\\\
-  turtleId = 19,\\\
-  z = 18,\\\
-  timestamp = 14.038611412048,\\\
+  x = 35,\\\
+  turtleId = 20,\\\
+  z = 20,\\\
+  timestamp = 16.878889083862,\\\
 }\""
